==> 1.TrendInDiagnosisPython_noplot.ipynb <==
#import libraries
import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
import seaborn as sns

#read file TableObe65plus and assign to object
icd = pd.read_csv('tblICD10.txt', dtype={0: str, 1: str, 2: str, 3: str, 4: str})

print(icd.info())

pd.set_option('display.width', 1000)
print(icd.head())

#specify columns 
icd.columns = ['patient', 'ICD', 'Hispanic', 'year']

print(icd.head())

#format the data
icd['ICD'] = icd['ICD'].str.replace('.*:', '', regex=True)
icd['ICD'] = icd['ICD'].str.replace('\..*', '', regex=True)
icd['Hispanic'] = icd['Hispanic'].str.replace('.*:', '', regex=True)

print(icd.head())

#estimate the frequency of each code per year
#icd_counts = icd.groupby('ICD', 'demog', 'year')['ICD'].count()
icd_counts = icd.groupby(['ICD', 'Hispanic', 'year']).size().reset_index(name='n')

print(icd_counts.head(20))

icd_totals = icd.groupby(['Hispanic', 'year'])['patient'].nunique().reset_index(name='total')

print(icd_totals.head(10))

frequencies = icd_counts.merge(icd_totals, on=['Hispanic', 'year'], how='left')
frequencies['percentage'] = round(frequencies['n'] / frequencies['total'] * 100, 2)
pd.set_option('display.max_rows', None) 
pd.set_option('display.max_columns', None) 
print(frequencies)

#export the dataframe to a txt file
frequencies.to_csv('frequencies.txt', index=False, header=True)
